use case configs
